docs(inital docs): The first official commit for Plus Ultra
This is the initial commit of the Team: Plus Ultra, The commit adds the
basic files that an MPLAB project includes.
It removes the README.md for the other files that were initially in
the project folder. It deletes the tutorial mymain.c file.
These files were removed because they were unnecessary for the 
build.