feat(Includes FreeRTOS): adds in FreeRTOS to project
This commit adds FREERTOS. to the project, once pushed this
commit should have no issues as it was able to build successfully.
The config file contains the change to allocate 1/2 as much memory to
the heap since I believe that it was too high since the FreeRTOS demo
which I got it from builds it with some unrealistic values.

Current memory usage:
Program: 78%
Data: 51%